Update aivo_get_info.m
Hi Tomi,

I started to work with Janne, we brainstormed a bit and thought how to solve it, the solution did not work out.
I completed the updating at home. I really tried to understand how to get the missing value, I understand the loop
is not elegant solution, but it happens ONLY if there are missing values, not often then.

Let me know what you think.