Add files via upload
The current version in the master returns a wrong order of values when a list of subject is passed.
In my original file with the sort resort idea I had before the value resorting a line with intersect(sub_ori,sub,'stable'),
that you discarded. It is necessary for the proper functioning of the function. Please check this new working and tested version.